ajout fichier signals_heredoc, gestion ctrl \ : affiche bien core dump quand une commande tourne (cat, wc) 18/02
